requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
PyMuPDF==1.24.10
python-docx==1.1.2
mammoth==1.8.0
pytesseract==0.3.13
pdf2image==1.17.0
